+++ NEWS.md
@@ -11,6 +11,11 @@ TODO:
   - e.g. independent test for different omics data and then combine results
 -->
 
+# clusterProfiler 4.19.3.001
+
++ `interpret()` now supports `task` parameter to specify the task: 'interpretation', 'annotation' and 'phenotyping' (2025-01-18, Sat)
++ `interpret()` supports `enrichResult`, `gseResult`, `compareClusterResult` and `list` of enrichment results (2025-01-18, Sat)
+
 # clusterProfiler 4.19.3
 
 + instead of packing KEGG cache data in the package, we now download it from <https://yulab-smu.top/clusterProfiler> (2025-12-15, Mon)